# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Election Management API
  version: 2.0.0
  description: API for managing users, voters, candidates, elections, and votes in an election management system.
servers:
  - url: http://localhost:3000/api
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Add a new user to the system (voter or admin).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationInput'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: User already exists or invalid input
        '500':
          description: Internal server error
  /auth/login:
    post:
      summary: Log in a user
      description: Authenticate a user and create a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                  enum: [voter, admin]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error
  /auth/logout:
    post:
      summary: Log out a user
      description: Destroy the user's session.
      responses:
        '200':
          description: Logged out successfully
        '500':
          description: Could not log out
  /auth/me:
    get:
      summary: Get current user
      description: Check if a user is authenticated and retrieve their data.
      responses:
        '200':
          description: User is authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  authenticated:
                    type: boolean
        '401':
          description: Not authenticated
...
  /elections:
    get:
      summary: Get all elections
      description: Retrieve a list of all ongoing and upcoming elections with their candidates.
      responses:
        '200':
          description: A list of elections.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Election'
        '500':
          description: Internal server error
    post:
      summary: Create a new election
      description: Add a new election to the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElectionInput'
      responses:
        '201':
          description: Election created successfully
  /elections/{id}:
    parameters:
      - name: id
        in: path
        description: ID of the election
        required: true
        schema:
          type: integer
    get:
      summary: Get an election by ID
      description: Retrieve a single election by its ID.
      responses:
        '200':
          description: Election found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Election'
        '404':
          description: Election not found
    put:
      summary: Update an election
      description: Update details of an existing election.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElectionInput'
      responses:
        '200':
          description: Election updated successfully
    delete:
      summary: Delete an election
      description: Remove an election from the system.
      responses:
        '200':
          description: Election deleted successfully
  /elections/{electionId}/vote-status/{userId}:
    get:
      summary: Check if a user has voted in an election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Voting status
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasVoted:
                    type: boolean
        '500':
          description: Internal server error
  /elections/{electionId}/vote:
    post:
      summary: Submit a vote
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                candidateId:
                  type: integer
                voterId:
                  type: integer
      responses:
        '200':
          description: Vote recorded successfully
        '400':
          description: Already voted
        '500':
          description: Internal server error
...
components:
  schemas:
    User:
      type: object
      properties:
        userid:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [voter, admin]
        gender:
          type: string
        date_of_birth:
          type: string
          format: date
...
    UserRegistrationInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [voter, admin]
        gender:
          type: string
        date_of_birth:
          type: string
          format: date
...
    Election:
      type: object
      properties:
        electionid:
          type: integer
        name:
          type: string
        type:
          type: string
        date:
          type: string
          format: date
        location_region:
          type: string
        status:
          type: string
          enum: [Upcoming, Ongoing, Completed, Canceled]
        candidates:
          type: array
          items:
            type: object
            properties:
              candidateid:
                type: integer
              name:
                type: string
              partyname:
                type: string
              symbol:
                type: string
    ElectionInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        date:
          type: string
          format: date
        location_region:
          type: string
        status:
          type: string
          enum: [Upcoming, Ongoing, Completed, Canceled]
...